router refactr, trying lag enhance
